removed static var for unknown packages
