+++ tests/tool/Integration.Tests/InterceptingKnownPackageLoader.cs
@@ -3,30 +3,27 @@
 
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
-using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.DotNet.UpgradeAssistant;
 using Microsoft.Extensions.Logging;
-using Xunit;
 
 namespace Integration.Tests
 {
     public class InterceptingKnownPackageLoader : IPackageLoader
     {
-        private static Dictionary<string, string> _unknownPackages = new Dictionary<string, string>();
-
+        private readonly UnknownPackages _unknownPackages;
         private readonly KnownPackages _packages;
         private readonly IPackageLoader _other;
         private readonly ILogger<InterceptingKnownPackageLoader> _logger;
 
-        public InterceptingKnownPackageLoader(KnownPackages packages, IPackageLoader other, ILogger<InterceptingKnownPackageLoader> logger)
+        public InterceptingKnownPackageLoader(KnownPackages packages, IPackageLoader other, UnknownPackages unknownPackages, ILogger<InterceptingKnownPackageLoader> logger)
         {
             _packages = packages ?? throw new ArgumentNullException(nameof(packages));
             _other = other ?? throw new ArgumentNullException(nameof(other));
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+            _unknownPackages = unknownPackages ?? throw new ArgumentNullException(nameof(unknownPackages));
         }
 
         public IEnumerable<string> PackageSources => _other.PackageSources;
@@ -71,28 +68,6 @@ namespace Integration.Tests
             }
 
             return latest ?? Array.Empty<NuGetReference>();
-        }
-
-        public static void ResetUnknownPackages()
-        {
-            _unknownPackages.Clear();
-        }
-
-        public static void AssertOnlyKnownPackagesWereReferenced(string actualDirectory)
-        {
-            if (!_unknownPackages.Keys.Any())
-            {
-                return;
-            }
-
-            var options = new JsonSerializerOptions
-            {
-                WriteIndented = true,
-            };
-            var uknownPackageStr = JsonSerializer.Serialize(_unknownPackages, options);
-            var outputFile = Path.Combine(actualDirectory, "UnknownPackages.json");
-            File.WriteAllText(outputFile, uknownPackageStr);
-            Assert.False(true, $"Integration tests tried to access NuGet.{Environment.NewLine}The list of packages not yet \"pinned\" has been written to:{Environment.NewLine}{outputFile}");
         }
     }
 }
