+++ tests/tool/Integration.Tests/UnknownPackages.cs
@@ -1,0 +1,11 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.Collections.Generic;
+
+namespace Integration.Tests
+{
+    public class UnknownPackages : Dictionary<string, string>
+    {
+    }
+}
